% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission_against_template.R
\name{check_submission_against_template}
\alias{check_submission_against_template}
\title{Check the consistency of an Rmd submission file against a reference template}
\usage{
check_submission_against_template(
  template_path = NULL,
  submission_path,
  template_number = NULL,
  reqvars_str = "Required variables",
  reqplot_str = "Required plots"
)
}
\arguments{
\item{template_path}{path to the template .Rmd file. Ignored if \code{template_number} is not NULL}

\item{submission_path}{path to the submission .Rmd file}

\item{template_number}{integer, indicating the number of the template you want to compare against.
If not \code{NULL}, the routine will ignore anything passed to \code{template_path} and retrieve the checkfile online.}

\item{reqvars_str}{Indicator string for definition of required variables}

\item{reqplot_str}{Indicator string for definition of required variables}
}
\value{
a data frame with the summary of the checks done, and specific indications
of any problem found.
}
\description{
This function runs the initial checks of a submission file to ensure that
\itemize{
\item all required code blocks are present
\item all required variables are generated
\item all required variables are of the correct type
\item no code block is returning any errors
\item no rogue calls to \code{install.packages} are done.
}
}
