% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_environments.R
\name{compare_environments}
\alias{compare_environments}
\title{Compare environments}
\usage{
compare_environments(sub_run, ref_run, template_info)
}
\arguments{
\item{sub_run}{submitted value}

\item{ref_run}{reference value}

\item{template_info}{parsed template_info obtained by running
lapply(parse_rmd_chunks(template_path),
parse_requirements).}
}
\value{
a list object with details of the check.
}
\description{
Compares results of running model Rmd vs. submitted Rmd, using template_info
to focus on required variables/plots.
Returns informative messages if there are differences
}
