% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_values.R
\name{compare_values}
\alias{compare_values}
\title{compare_values:}
\usage{
compare_values(sub_val, ref_val, tol = 1e-06)
}
\arguments{
\item{sub_val}{submitted value}

\item{ref_val}{reference value}

\item{tol}{tolerance for numerical comparisons}
}
\value{
a list objecti with fields match (logical), detail (string)
}
\description{
Compares values and returns informative messages if there are differences
}
