% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_SCEMChecker.R
\name{summary.SCEMChecker}
\alias{summary.SCEMChecker}
\title{summary.SCEMChecker}
\usage{
\method{summary}{SCEMChecker}(object, ...)
}
\arguments{
\item{object}{object of class \emph{SCEMChecker}.}

\item{...}{other parameters to be passed down to specific
summary functions (currently unused)}
}
\value{
Nothing (called for side effects)
}
\description{
S3 method for summarizing \emph{SCEMChecker} objects output by \code{\link[=check_submission_against_template]{check_submission_against_template()}}).
}
