% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cw_datasets.R
\name{get_cw_datasets}
\alias{get_cw_datasets}
\title{Get datasets for the SCEM coursework}
\usage{
get_cw_datasets(
  seed,
  N,
  maindata = NULL,
  save.folder = NULL,
  options = list(lmin = 2000, lmax = 4000, class.column = "Class", class.neg = 0,
    class.pos = 1)
)
}
\arguments{
\item{seed}{integer used for dataset generation. Unless otherwise stated,
it should be your student ID.}

\item{N}{the desired number of datasets to be generated.}

\item{maindata}{a data.frame containing the source dataset from which further
datasets will be instantiated. This will commonly be provided by the unit
director.}

\item{save.folder}{path to the folder where the resulting data should be saved.
Use \code{NULL} if you don't wish the data to be saved to disk.}

\item{options}{further options to be passed to the dataset generator. If needed,
these will be specified by the unit director.}
}
\value{
a named list object containing \code{N} datasets generated.
}
\description{
This function is used to instantiate datasets for activities related to the
EMATM0061 - Statistical Computing and Empirical Methods unit at the
University of Bristol.
}
