% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmd_chunks.R
\name{run_rmd_chunks}
\alias{run_rmd_chunks}
\title{Run an Rmd and capture outputs}
\usage{
run_rmd_chunks(parsed, remove_illegal_installs = TRUE, env = NULL)
}
\arguments{
\item{parsed}{a parsed Rmd file}

\item{remove_illegal_installs}{logical, should calls to install.packages() be
removed outside the "Setup" chunk?}

\item{env}{environment in which to evaluate the chunks. If \code{NULL},  a new environment is created.

@returns a list object with:
\itemize{
\item env: environment with variables created
\item chunks: named list per chunk with fields: chunk, present,
error, plot (recordedplot or NULL), plot_generated (logical),
illegal_install (logical)
}

@export}
}
\description{
Run an Rmd and capture outputs
}
